finalize cleanup and remove grpc-adapter
