--- components/reactions/grpc-adaptive/CHANGELOG.md
@@ -1,7 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-<!-- generated by git-cliff -->
