--- components/reactions/grpc-adaptive/Cargo.toml
@@ -1,48 +1,0 @@
-# Copyright 2025 The Drasi Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-[package]
-name = "drasi-reaction-grpc-adaptive"
-version = "0.1.2"
-edition = "2021"
-authors = ["Drasi Project"]
-description = "gRPC Adaptive reaction plugin for Drasi"
-license = "Apache-2.0"
-repository = "https://github.com/drasi-project/drasi-core"
-keywords = ["drasi", "plugin", "reaction", "grpc", "adaptive"]
-categories = ["network-programming"]
-
-[lints]
-workspace = true
-
-[dependencies]
-drasi-lib.workspace = true
-drasi-reaction-grpc.workspace = true
-anyhow = "1.0"
-async-trait = "0.1"
-log = "0.4"
-tonic = "0.11"
-prost = "0.12"
-tokio = { version = "1.0", features = ["full"] }
-tokio-stream = { version = "0.1", features = ["sync"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-chrono = "0.4"
-uuid = { version = "1.0", features = ["v4"] }
-prost-types = "0.12"
-ordered-float = "3.0"
-
-[build-dependencies]
-tonic-build = "0.11"
