+++ components/reactions/grpc/src/config.rs
@@ -13,6 +13,67 @@
 // limitations under the License.
 
 //! Configuration types for gRPC reactions.
+//! This reaction implements gRPC communication for Drasi.
+//! Supports fixed and adaptive batching.
+//!
+//! Fixed batching example:
+//!
+//! ```rust,ignore
+//! use drasi_lib::config::{ReactionConfig, ReactionSpecificConfig};
+//! use drasi_lib::reactions::grpc::GrpcReactionConfig;
+//! use drasi_lib::reactions::common::AdaptiveBatchConfig;
+//! use std::collections::HashMap;
+//!
+//! let grpc_config = GrpcReactionConfig {
+//!     endpoint: "grpc://localhost:50052".to_string(),
+//!     timeout_ms: 5000,
+//!     batch_size: 100,
+//!     batch_flush_timeout_ms: 1000,
+//!     max_retries: 3,
+//!     connection_retry_attempts: 5,
+//!     initial_connection_timeout_ms: 10000,
+//!     metadata: HashMap::new(),
+//!     adaptive_enable: false,
+//!     adaptive: Default::default(),
+//! };
+//! let reaction_config = ReactionConfig {
+//!     id: "my-grpc-reaction".to_string(),
+//!     queries: vec!["query1".to_string()],
+//!     auto_start: true,
+//!     config: ReactionSpecificConfig::Grpc(grpc_config),
+//!     priority_queue_capacity: Some(500),
+//! };
+//! ```
+//! Adaptive batching example:
+//!
+//! ```rust,ignore
+//! use drasi_lib::config::{ReactionConfig, ReactionSpecificConfig};
+//! use drasi_lib::reactions::grpc::GrpcReactionConfig;
+//! use drasi_lib::reactions::common::AdaptiveBatchConfig;
+//! use std::collections::HashMap;
+//!
+//! let reaction_config = ReactionConfig {
+//!     id: "high-throughput".to_string(),
+//!     queries: vec!["event-stream".to_string()],
+//!     auto_start: true,
+//!     config: ReactionSpecificConfig::Grpc(GrpcReactionConfig {
+//!         endpoint: "grpc://event-processor:9090".to_string(),
+//!         timeout_ms: 15000,
+//!         max_retries: 5,
+//!         connection_retry_attempts: 5,
+//!         initial_connection_timeout_ms: 10000,
+//!         metadata: HashMap::new(),
+//!         adaptive_enable: true,
+//!         adaptive: AdaptiveBatchConfig {
+//!             adaptive_min_batch_size: 50,
+//!             adaptive_max_batch_size: 2000,
+//!             adaptive_window_size: 100,  // 10 seconds
+//!             adaptive_batch_timeout_ms: 500,
+//!         },
+//!     }),
+//!     priority_queue_capacity: None,
+//! };
+//! ```
 
 use drasi_lib::reactions::common::AdaptiveBatchConfig;
 use serde::{Deserialize, Serialize};
